Fixed a length bug with hyperlink blocks.
Text block render distance is now based on text size. The default text size (1) now has a render distance of 40 (was 64), with the maximum render distance being 400 blocks at size 34. This should help performance and usability!

Added a catch for a certain crash that I can't find the cause of in the TextBlockEditScreen. Will still fail, but will no longer crash the game.